+++ src/quin_huffman.c
@@ -17,7 +17,7 @@ int quinary_node_cmp(const void *node_1, const void *node_2) {
 }
 
 int quinary_huffman_pack(FILE *in, FILE *out) {
-    size_t cnt[256] = {0};
+    uint64_t cnt[256] = {0};
     unsigned char value;
     while (fread(&value, sizeof(value), 1, in) == sizeof(value)) {
         ++cnt[value];
@@ -28,7 +28,7 @@ int quinary_huffman_pack(FILE *in, FILE *out) {
     // Quinary huffman needs node counts satisfying c % 4 == 1.
     // 257 % 4 == 1, so at most there will be 257 nodes.
     quin_huffman_node_t *huffman_nodes[257] = {0};
-    size_t node_count = 0;
+    uint64_t node_count = 0;
     for (int i = 0; i != 256; ++i) {
         if (cnt[i]) {
             huffman_nodes[node_count] = malloc(sizeof(quin_huffman_node_t));
@@ -106,7 +106,7 @@ int quinary_huffman_pack(FILE *in, FILE *out) {
     }
 #endif
 
-    size_t dict_count = 0;
+    uint64_t dict_count = 0;
     quin_huffman_head_dict_t head_dict[256];
     for (unsigned int i = 0; i != 256; ++i) {
         if (dict[i].bit_length != 0) {
@@ -119,7 +119,7 @@ int quinary_huffman_pack(FILE *in, FILE *out) {
     // Write magic head.
     fwrite(&quin_huffman_head_magic, sizeof(quin_huffman_head_magic), 1, out);
     quin_huffman_head_t head = {
-        .data_size = (size_t)fsize, .dict_count = dict_count, .reserved = {0}};
+        .data_size = (uint64_t)fsize, .dict_count = dict_count, .reserved = {0}};
     // Write head.
     fwrite(&head, sizeof(head), 1, out);
     // Write dict.
@@ -211,9 +211,9 @@ void quinary_huffman_tree_walk(quin_huffman_node_t *const root,
     }
 }
 
-int quinary_huffman_get_data_size(FILE *in, size_t *out) {
-    u_int32_t magic;
-    fread(&magic, sizeof(u_int32_t), 1, in);
+int quinary_huffman_get_data_size(FILE *in, uint64_t *out) {
+    uint32_t magic;
+    fread(&magic, sizeof(uint32_t), 1, in);
     if (magic != quin_huffman_head_magic) {
         return -1;
     }
@@ -226,8 +226,8 @@ int quinary_huffman_get_data_size(FILE *in, size_t *out) {
 
 int quinary_huffman_unpack(FILE *in, FILE *out) {
 
-    u_int32_t magic;
-    fread(&magic, sizeof(u_int32_t), 1, in);
+    uint32_t magic;
+    fread(&magic, sizeof(uint32_t), 1, in);
     if (magic != quin_huffman_head_magic) {
         return -1;
     }
@@ -237,7 +237,7 @@ int quinary_huffman_unpack(FILE *in, FILE *out) {
         malloc(sizeof(quin_huffman_head_dict_t) * head.dict_count);
     fread(dict, sizeof(quin_huffman_head_dict_t), head.dict_count, in);
 #ifdef DEBUG_OUTPUT
-    for (size_t i = 0; i != head.dict_count; ++i) {
+    for (uint64_t i = 0; i != head.dict_count; ++i) {
         printf("in-file code for: 0x%x (%c): ", dict[i].byte, dict[i].byte);
         for (int j = (int)dict[i].code.bit_length - 1; j >= 0; --j) {
             printf("%d", dict[i].code.qbits[j].bit);
@@ -267,7 +267,7 @@ int quinary_huffman_unpack(FILE *in, FILE *out) {
         }
     }
 #endif
-    size_t decoded_byte_count = 0;
+    uint64_t decoded_byte_count = 0;
     unsigned char byte = 0;
     quin_huffman_node_t *position = root;
     // Walk a step when every three bits are read into tbit_temp.
@@ -322,13 +322,13 @@ END_LOOP:
 
 void quinary_huffman_tree_build(quin_huffman_node_t *const root,
                                 quin_huffman_char_t code,
-                                quin_huffman_head_dict_t *dict, size_t n,
+                                quin_huffman_head_dict_t *dict, uint64_t n,
                                 quin_huffman_head_dummy_list_t *dummy_list) {
 
-    for (size_t i = 0; i != n; ++i) {
+    for (uint64_t i = 0; i != n; ++i) {
         if (code.bit_length == dict[i].code.bit_length) {
-            u_int16_t bc = code.bit_length;
-            u_int16_t j = 0;
+            uint16_t bc = code.bit_length;
+            uint16_t j = 0;
             for (j = 0;
                  j != bc && code.qbits[j].bit == dict[i].code.qbits[j].bit; ++j)
                 ;
@@ -349,8 +349,8 @@ void quinary_huffman_tree_build(quin_huffman_node_t *const root,
     for (int i = 0; i != dummy_list->count; ++i) {
         quin_huffman_char_t *dummy = &dummy_list->codes[i];
         if (dummy != NULL && code.bit_length == dummy->bit_length) {
-            u_int16_t i;
-            u_int16_t bc = dummy->bit_length;
+            uint16_t i;
+            uint16_t bc = dummy->bit_length;
             for (i = 0; i != bc && code.qbits[i].bit == dummy->qbits[i].bit;
                  ++i)
                 ;
